LayoutTests/compositing/plugins/pdf/pdf-plugin-printing.html

<!DOCTYPE html> <!-- webkit-test-runner [ UnifiedPDFEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Basic testing for automatic PDF printing");
jsTestIsAsync = true;
let receivedBeforePrintEvent = false;

addEventListener("beforeprint", () => {
    debug("* Received beforeprint event");
    receivedBeforePrintEvent = true;
});

addEventListener("afterprint", () => {
    debug("* Received afterprint event");
    shouldBeTrue("receivedBeforePrintEvent");
    finishJSTest();
});
</script>
<embed src="../../../resources/auto-print.pdf"></embed>
</body>
</html>
